optee: multiplex tee interrupt events on optee async notif irq
Implements an irqchip in optee driver for the interrupt events notified
from OP-TEE world to the Linux OS. Optee registers up to 1 interrupt
controller and identifies each line with a line number from 0 to
UINT16_MAX. The identifiers and meaning of the interrupt line number
are specific to the platform and shall be found in the OP-TEE platform
documentation

There already exist an optee asynchronous notification mechanism using
an irq for OP-TEE to signal its optee driver in Linux of a pending
interrupt event. The implementation binds each event to the awaking of
a thread waiting to call the TEE. The interrupt notification service
added by this change allows TEE to send interrupt signals to Linux kernel
on secure event occurrences which end consumer is a Linux sub-system.

When optee driver initializes, TEE tells whether it supports interrupt
notification services or not.

OP-TEE SMC ABI defines 4 new SMC function IDs for non-secure world to
manage interrupt events:
- OPTEE_SMC_GET_NOTIF_ITR allows non-secure world to retrieve pending
  interrupts by grapes up to 5 lines. For efficiency, the function also
  reports whether there are pending async values targeting suspended
  threaded sequences execution and whether TEE has background threaded
  work to do.
- OPTEE_SMC_NOTIF_ITR_SET_MASK allows Linux irq consumer to mask/unmask
  the irq line using a fastcall SMC (uninterruptible) service.
- OPTEE_SMC_NOTIF_ITR_SET_STATE allows Linux irq consumer to enable and
  disable the irq line using a yielded TEE call.
- OPTEE_SMC_NOTIF_ITR_SET_WAKEUP allows Linux irq consumer to enable
  and disable the wakeup from low power capability of the interrupt.

This change was posted to the LKML (patch v4 [1]). A patch v5 [2] exists
and is equivalent to this change but targets kernel v6.3-rc4 and later
to resolve rebase conflicts.

Link: [1] https://lore.kernel.org/lkml/20230317171003.974801-2-etienne.carriere@linaro.org/
Link: [2] https://lore.kernel.org/lkml/20230506073235.2770292-2-etienne.carriere@linaro.org/
Co-developed-by: Pascal Paillet <p.paillet@foss.st.com>
Signed-off-by: Pascal Paillet <p.paillet@foss.st.com>
Co-developed-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>