arm64: dt: OP-TEE for FVP Base RevC
Configures FVP Base RevC with OP-TEE.

With commit 1bf41f5d9b0d ("fvp: add Trusted Services support") in
https://github.com/OP-TEE/build there is support for building running
with the FVP Base RevC model. However, some updates to the DTB passed
to the kernel is missing for the OP-TEE driver to be probed. So fix that
by adding an OP-TEE node in the DTS.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>